\name{Q.score}
\alias{Q.score}

\title{Q score for foraging efficiency.

}
\description{
Q score is an indicator of visual search efficiency which accounts for speed and accuracy (Geldmacher & Hills, 1997).
}
\usage{
Q.score(n.collected.targets,n.total.targets,rt)
}

\arguments{
  \item{n.collected.targets}{
A numeric value with the number of targets collected in a trial.
}
\item{n.total.targets}{
A numeric value with the total number of targets present in a trial.
}
\item{rt}{
A numeric value with the total time spent on a display or patch.
}
}
\value{
A numeric value with the Q score.
}
\references{
Geldmacher, D. S. & Hills, E. C. (1997) Effect of stimulus number, target-to-distractor ratio, and motor speed  visual spatial search quality following traumatic brain injury. \emph{Brain Injury, 11}(1), 59-66.
}
\author{
Marcos Bella-Fernández
}
